requirements.txt
numpy
opencv-python
streamlit
pandas
pillow
tensorflow
mediapipe